requirements.txt
Django==4.0.2
django-import-export==2.8.0
